requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
python-dotenv>=1.0.1
pydantic>=2.6.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
openpyxl>=3.1.2

strawberry-graphql[fastapi]==0.210.0
graphql-core>=3.2.3

# Dash 대시보드 추가 부분 👇
dash>=2.17.0
plotly>=5.22.0

